Restructure app to work with routers
