+++ src/index.js
@@ -3,7 +3,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './styles/index.scss';
-import TodoApp from './TodoApp';
+import TodoApp from './routes/Home';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
